added graph types
